--- requirements.txt
@@ -1,4 +1,0 @@
-Faker==15.3.*
-GitPython==3.1.*
-coloredlogs==15.0.*
-networkx==2.8.*
