+++ satellite/__init__.py
@@ -1,0 +1,3 @@
+from satellite.main import cli
+
+__all__ = ["cli"]
