+++ satellite/_log.py
@@ -1,0 +1,5 @@
+import logging
+import coloredlogs
+
+logger = logging.getLogger(__name__)
+coloredlogs.install(level=logging.INFO, logger=logger)
